+++ src/internal/interpreter/map-interpreter.test.ts
@@ -3,6 +3,7 @@ import { parseMap, Source } from '@superfaceai/parser';
 import { getLocal } from 'mockttp';
 
 import { CrossFetch } from '../../lib/fetch';
+import { ServicesSelector } from '../../lib/services';
 import { MapInterpreter } from './map-interpreter';
 
 const mockServer = getLocal();
@@ -19,11 +20,10 @@ const parseMapFromSource = (source: string) =>
   );
 
 describe('MapInterpreter', () => {
-  let serviceBaseUrl: string;
-
+  let mockServicesSelector: ServicesSelector;
   beforeEach(async () => {
     await mockServer.start();
-    serviceBaseUrl = mockServer.url;
+    mockServicesSelector = ServicesSelector.withDefaultUrl(mockServer.url);
   });
 
   afterEach(async () => {
@@ -35,6 +35,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'testCase',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -53,6 +54,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -72,6 +74,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -89,6 +92,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -111,6 +115,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -134,6 +139,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -167,7 +173,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -198,7 +204,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { page: '2' },
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -233,7 +239,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { page: 2 },
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -270,7 +276,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -308,7 +314,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -354,7 +360,7 @@ describe('MapInterpreter', () => {
             password: 'password',
           },
         ],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -390,7 +396,7 @@ describe('MapInterpreter', () => {
             token: 'SuperSecret',
           },
         ],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -427,7 +433,7 @@ describe('MapInterpreter', () => {
             apikey: 'SuperSecret',
           },
         ],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -465,7 +471,7 @@ describe('MapInterpreter', () => {
             apikey: 'SuperSecret',
           },
         ],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -507,7 +513,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'testCase',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -541,7 +547,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'testCase',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -570,6 +576,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'testCase',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -600,6 +607,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { condition: true },
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -608,6 +616,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { condition: false },
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -622,6 +631,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -644,7 +654,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -691,6 +701,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -734,7 +745,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -765,6 +776,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -792,6 +804,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -819,6 +832,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -845,6 +859,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -872,6 +887,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -900,6 +916,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -926,6 +943,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -952,7 +970,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { test: 'twelve' },
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -979,8 +997,8 @@ describe('MapInterpreter', () => {
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
-        serviceBaseUrl: baseUrl,
         security: [],
+        services: ServicesSelector.withDefaultUrl(baseUrl),
       },
       { fetchInstance }
     );
@@ -1004,7 +1022,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -1037,7 +1055,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
         parameters: {
           message: 'nice!',
         },
@@ -1068,7 +1086,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         parameters: { path: 'twelve' },
         security: [],
-        serviceBaseUrl,
+        services: mockServicesSelector,
       },
       { fetchInstance }
     );
@@ -1094,12 +1112,70 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         parameters: { path: 'twelve' },
         security: [],
-        serviceBaseUrl: `${serviceBaseUrl}/{path}`,
+        services: ServicesSelector.withDefaultUrl(
+          `${mockServicesSelector.getUrl()!}/{path}`
+        ),
       },
       { fetchInstance }
     );
     const result = await interpreter.perform(ast);
     expect(result.isOk() && result.value).toEqual({ result: 12 });
+  });
+
+  it('should correctly select service', async () => {
+    const urlOne = '/one/something';
+    await mockServer.get(urlOne).thenJson(200, { data: 1 });
+
+    const urlTwo = '/two/something';
+    await mockServer.get(urlTwo).thenJson(200, { data: 2 });
+
+    const urlThree = '/three/something';
+    await mockServer.get(urlThree).thenJson(200, { data: 3 });
+
+    const ast = parseMapFromSource(`
+      map Test {
+        value = 0
+
+        http GET default "/something" {
+          response 200 "application/json" {
+            value = value + body.data
+          }
+        }
+
+        http GET "two" "/something" {
+          response 200 "application/json" {
+            value = value + body.data
+          }
+        }
+
+        http GET "three" "/something" {
+          response 200 "application/json" {
+            value = value + body.data
+          }
+        }
+
+        map result {
+          result = value
+        }
+      }`);
+
+    const interpreter = new MapInterpreter(
+      {
+        usecase: 'Test',
+        security: [],
+        services: new ServicesSelector(
+          [
+            { id: 'one', baseUrl: `${mockServicesSelector.getUrl()!}/one` },
+            { id: 'two', baseUrl: `${mockServicesSelector.getUrl()!}/two` },
+            { id: 'three', baseUrl: `${mockServicesSelector.getUrl()!}/three` },
+          ],
+          'one'
+        ),
+      },
+      { fetchInstance }
+    );
+    const result = await interpreter.perform(ast);
+    expect(result.isOk() && result.value).toEqual({ result: 6 });
   });
 
   it('should correctly return error from operation', async () => {
@@ -1117,6 +1193,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -1141,6 +1218,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -1168,6 +1246,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -1198,6 +1277,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );
@@ -1222,6 +1302,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        services: ServicesSelector.withDefaultUrl(''),
       },
       { fetchInstance }
     );