+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added support for service selection in http calls
+
 ## [1.1.0] - 2021-12-22
 ### Added
 - Pass integration parameters in `perform`